refactor(Components): Create header subfolder
