+++ src/components/header-footer/CartWidget.js
@@ -1,9 +1,9 @@
 import { useContext } from "react";
-import { CartContext } from "../contexts/CartContext";
+import { CartContext } from "../../contexts/CartContext";
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faCartShopping } from '@fortawesome/free-solid-svg-icons';
 import StyledNavLink from "./StyledNavLink";
-import { CartWidgetContainer } from "./StyledComponents";
+import { CartWidgetContainer } from "../StyledComponents";
 
 export default function CartWidget() {
     const { cart } = useContext(CartContext);
