+++ src/components/header-footer/HeaderFooter.js
@@ -1,7 +1,7 @@
 import { useContext } from "react";
-import { ThemeContext } from "../contexts/ThemeContext";
+import { ThemeContext } from "../../contexts/ThemeContext";
 import { Outlet } from "react-router-dom";
-import { HeaderContainer, PageContainer } from "./StyledComponents";
+import { HeaderContainer, PageContainer } from "../StyledComponents";
 import CartWidget from "./CartWidget";
 import NavBar from "./NavBar";
