+++ src/components/header-footer/StyledNavLink.js
@@ -1,7 +1,7 @@
 import { useContext } from "react";
-import { ThemeContext } from "../contexts/ThemeContext";
+import { ThemeContext } from "../../contexts/ThemeContext";
 import { NavLink } from "react-router-dom";
-import { NavLinkText } from "./StyledComponents";
+import { NavLinkText } from "../StyledComponents";
 
 export default function StyledNavLink({ page, text }) {
     const { theme } = useContext(ThemeContext);
